* onechiporenko/master: (101 commits)
  Add sort function per column (#308)
  v2.5.0
  #304 Allow to set colspan for table header cells
  #303 Add a "summary"-row for grouped rows
  #302 Add flag `simple` to the columns
  v2.4.0
  Split `_createColumn`
  Fix tests for ember@3.0+
  Fix 2.8.0 build and linting
  Move to ember-cli@3.0.0
  Fix `data` binding to the columns
  expose records count via publicAPI (#301)
  Allow CSS class to be set for table header. (#299)
  Added selectedItems, visibleProcessedColumns and data to the yielded
block of table-footer (#298)
  Update build commands
  v2.3.1
  Fix travis
  #295 Toggle all selection is broken
  Add info about slack-channel
  Add examples from ember-twiddle
  ...

# Conflicts:
#	addon/components/models-table.js
#	tests/integration/components/models-table-test.js

+++ gh-pages.sh
@@ -1,15 +1,21 @@
 #!/bin/bash
 
-git branch -D gh-pages
-git push origin --delete gh-pages
-git checkout -b gh-pages
-ember build --environment production
-git rm -rf app addon config tests
-git rm -rf Brocfile.js bower.json package.json testem.json
+npm run build:gh-pages
+git checkout gh-pages
+# from https://stackoverflow.com/questions/37890510/bash-script-to-check-if-the-current-git-branch-x
+BRANCH=$(git rev-parse --abbrev-ref HEAD)
+if [[ "$BRANCH" != "gh-pages" ]]; then
+  echo 'Aborting script';
+  exit 1;
+fi
+git rm -rf app addon config tests blueprints
+git rm -rf Brocfile.js bower.json package.json package-lock.json testem.json
 git rm -rf .bowerrc .editorconfig .jshintrc .travis.yml
-mv dist/* .
+git rm -rf v.2
+mkdir ./v.2
+mv dist/* ./v.2
 rm -rf dist
-git add .
-git commit -m "Publishing to github pages"
+git add v.2
+git commit -m "Pushing v.2 to github pages"
 git push origin gh-pages
 git checkout master
